// libs/shared/core/src/Colors/colors.type.ts
export type botColors = {
  bot: {
    blue: { hex: string; rgb: string };
    pink: { hex: string; rgb: string };
  };
  company: {
    gold: { hex: string; rgb: string };
    silver: { hex: string; rgb: string };
    black: { hex: string; rgb: string };
  };
  system: {
    critical: { hex: string; rgb: string };
    error: { hex: string; rgb: string };
    warning: { hex: string; rgb: string };
    info: { hex: string; rgb: string };
    success: { hex: string; rgb: string };
  };
};